/POST/salir
eliminacion del filtro
