+++ core/src/main/java/gt/edu/umes/broker/core/system/EncryptionUtil.java
@@ -1,0 +1,10 @@
+package gt.edu.umes.broker.core.system;
+
+public class EncryptionUtil {
+    private static final String ALGORITHM = "AES/GCM/NoPadding";
+    private static final int GCM_IV_LENGTH = 12;
+    private static final int GCM_TAG_LENGTH = 128;
+
+
+
+}
